requirements.txt
Flask
Flask-WTF
Flask-Cors
Flask-Login
Flask-Bcrypt
Flask-Limiter
Flask-Caching
Flask-SQLAlchemy
alembic==1.13.1
black
email-validator
gdown==5.2.0
gradio==5.35.0
gradio_client==1.10.4
gunicorn
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
pandas==2.2.3
pandasql==0.7.3
Pillow==10.0.0
pydantic
pytest
python-dotenv
requests
SQLAlchemy==2.0.35
timm==1.0.19
torch==2.5.0
torchvision==0.20.0
transformers==4.51.3
ultralytics==8.3.169
ultralytics-thop==2.0.14
Werkzeug==2.3.7
